requirements.txt
gevent>=21.12.0
requests>=2.26.0
PyYAML>=6.0
jinja2>=3.0.0
croniter>=1.3.0
Babel>=2.9.0
python-dateutil>=2.8.0
celery>=5.2.0
redis>=4.0.0
tornado>=6.1.0
pymysql>=1.0.0
sqlalchemy>=1.4.0
bcrypt>=3.2.0
pyjwt>=2.4.0
cryptography>=3.4.0
